pyproject.toml
[project]
name = "synapsteward-climatecore"
version = "0.1.0"
description = "Control unit for environmental sensors and actuators in SynapSteward"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llm>=0.22",
    "llm-mlx>=0.3",
    "mypy>=1.15.0",
    "nats-py>=2.9.0",
    "pydantic-settings>=2.7.1",
    "pydantic>=2.10.6",
    "retry>=0.9.2",
]
